Set safe.directory in git config
Some nuget packages, like GitVersion, run git commands. Because the
dotnet application is being compiled inside of a docker image, we need
to specify the `safe.directory` global git configuration so that the
`/github/workspace` git repository path does not error out when git
commands are run inside of it.